requirements.txt
aniso8601==10.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.3.3
cycler==0.12.1
filelock==3.18.0
Flask==3.1.0
flask-restx==1.3.0
fonttools==4.59.0
fsspec==2025.7.0
ftfy==6.3.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflect==7.5.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.5
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
nltk==3.9.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.24.0
ruff==0.11.8
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
soupsieve==2.7
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typeguard==4.4.4
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
youtube-dl==2021.12.17
zipp==3.21.0

clip~=1.0